requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
SQLAlchemy==2.0.40
pydantic-settings==2.9.1
python-docx==1.1.2
httpx==0.28.1
